fix: default a format-less number schema to double
`NumberSchema` normalized a missing or unknown format to `float`, so a
`type: number` property without format was generated as `java.lang.Float`.
JSON numbers are double-precision, hence such a property silently lost
precision on a round-trip. Default to `double` instead; an explicitly
declared `float` still yields `Float`.

The `decimal` mapping of the #381 example converts from the normalized
type of an unknown format, which is `Double` now.

Issue: #396

+++ plugin-build/plugin/src/main/java/com/github/muehmar/gradle/openapi/generator/model/schema/NumberSchema.java
@@ -68,10 +68,10 @@ public class NumberSchema implements OpenApiSchema {
         .map(
             formatValue ->
                 NumericType.ofFormatAndValue(
-                    NumericType.Format.parseString(formatValue).orElse(NumericType.Format.FLOAT),
+                    NumericType.Format.parseString(formatValue).orElse(NumericType.Format.DOUBLE),
                     formatValue,
                     nullability))
-        .orElseGet(() -> NumericType.noFormat(NumericType.Format.FLOAT, nullability))
+        .orElseGet(() -> NumericType.noFormat(NumericType.Format.DOUBLE, nullability))
         .withConstraints(constraints);
   }
 }
